+++ src/main/java/com/transport/management/modules/ruta/RutaEntity.java
@@ -1,4 +1,4 @@
-package com.transport.management.entities;
+package com.transport.management.modules.ruta;
 
 import com.transport.management.utils.abtractBase.BaseEntity;
