+++ src/main/java/com/transport/management/modules/ruta/RutaRepository.java
@@ -1,6 +1,5 @@
-package com.transport.management.repositories;
+package com.transport.management.modules.ruta;
 
-import com.transport.management.entities.RutaEntity;
 import com.transport.management.utils.abtractBase.BaseRepository;
 
 public interface RutaRepository extends BaseRepository<RutaEntity> {}
