+++ src/main/java/com/transport/management/modules/ruta/RutaService.java
@@ -1,8 +1,7 @@
-package com.transport.management.services;
+package com.transport.management.modules.ruta;
 
 import org.springframework.stereotype.Service;
 
-import com.transport.management.entities.RutaEntity;
 import com.transport.management.utils.abtractBase.BaseService;
 
 @Service
