+++ src/main/java/com/transport/management/modules/viaje/ViajeEntity.java
@@ -4,10 +4,10 @@ import java.time.LocalDateTime;
 import java.util.Random;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.transport.management.entities.RutaEntity;
 import com.transport.management.enums.StatusViajeEnum;
 import com.transport.management.modules.cliente.ClienteEntity;
 import com.transport.management.modules.conductor.ConductorEntity;
+import com.transport.management.modules.ruta.RutaEntity;
 import com.transport.management.utils.abtractBase.BaseEntity;
 
 import jakarta.persistence.Column;
